Fixed some issues caused by merge. Prices broken

+++ WegmansClicker/Assets/Scripts/CatalogItem.cs
@@ -43,6 +43,6 @@ public class CatalogItem : MonoBehaviour
 
     public void MakePurchase()
     {
-        category.increaseValue((int)food.price);
+        category.increaseValue(food.price);
     }
 }
